Add atomic compound operations
They allow performing a sequence of operations guaranteed to be
sequential. Even if a concurrent (literally parallel) operation
is produced, it will never occur within the sequence of operations
defined as atomic.

It is implemented by having those operations share the same
Lamport timestamp, except for the last integer, which gets incremented.